+++ R/nearmiss_impl.R
@@ -1,3 +1,61 @@
+#' Remove Points Near Other Classes
+#'
+#' Generates synthetic positive instances using nearmiss algorithm.
+#'
+#' @inheritParams step_nearmiss
+#' @param df data.frame or tibble. Must have 1 factor variable and remaining
+#'  numeric variables.
+#' @param var Character, name of variable containing factor variable.
+#' @param k An integer. Number of nearest neighbor that are used
+#'  to generate the new examples of the minority class.
+#'
+#' @return A data.frame or tibble, depending on type of `df`.
+#' @export
+#'
+#' @details
+#' All columns used in this function must be numeric with no missing data.
+#'
+#' @references Inderjeet Mani and I Zhang. knn approach to unbalanced data
+#' distributions: a case study involving information extraction. In Proceedings
+#' of workshop on learning from imbalanced datasets, 2003.
+#'
+#' @family Direct Implementations
+#'
+#' @examples
+#' circle_numeric <- circle_example[, c("x", "y", "class")]
+#'
+#' res <- nearmiss(circle_numeric, var = "class")
+#'
+#' res <- nearmiss(circle_numeric, var = "class", k = 10)
+#'
+#' res <- nearmiss(circle_numeric, var = "class", under_ratio = 1.5)
+nearmiss <- function(df, var, k = 5, under_ratio = 1) {
+  if (length(var) != 1) {
+    rlang::abort("Please select a single factor variable for `var`.")
+  }
+
+  var <- rlang::arg_match(var, colnames(df))
+
+  if (!(is.factor(df[[var]]) | is.character(df[[var]]))) {
+    rlang::abort(glue("`{var}` should be a factor or character variable."))
+  }
+
+  if (length(k) != 1) {
+    rlang::abort("`k` must be length 1.")
+  }
+
+  if (k < 1) {
+    rlang::abort("`k` must be non-negative.")
+  }
+
+  predictors <- setdiff(colnames(df), var)
+
+  check_numeric(df[, predictors])
+  check_na(select(df, -all_of(var)), "nearmiss")
+
+  nearmiss_impl(df, var, ignore_vars = character(), k, under_ratio)
+}
+
 nearmiss_impl <- function(df, var, ignore_vars, k = 5, under_ratio = 1) {
   classes <- downsample_count(df, var, under_ratio)