+++ NEWS.md
@@ -6,6 +6,7 @@
 * All recipe steps now officially support empty selections to be more aligned with dplyr and other packages that use tidyselect (#55)
 * `step_smotenc()` have been added to implement SMOTENC which can handle categorical as well as numerical values. Thanks to @RobertGregg (#82)
 * Fix bug where wrong ordering of columns caused error in `smote()`. (#76)
+* export `nearmiss()` functions to users.
 
 # themis 0.1.4